==> notebooks/01_preprocessing.ipynb <==
## Обработка данных

import pandas as pd
import os

pd.set_option('display.max_colwidth', None)
pd.set_option('display.max_columns', None)
pd.set_option('display.width', None)

os.chdir('../')

df = pd.read_excel('data/orig/test.xlsx')
df.head()

# Проверяем на пустые значения в столбцах, если вдруг надо обработать их
df.isnull().sum()

Так как цифры встречаются очень часто и не несут смысловой нагрузки, можем обработать их, как и знаки препинания

def apply_and_check(dataframe, func, column, target_column=None, **kwargs):
    if target_column is None:
        target_column = column

    total_before = dataframe[column].str.len().sum()
    dataframe[target_column] = dataframe[column].apply(func, **kwargs)

    total_after = dataframe[target_column].str.len().sum()
    total_removed = total_before - total_after
    
    print(f'Символов до: {total_before:,}, '
            f'после: {total_after:,}')
    print(f'Удалено символов: {total_removed:,} ({total_removed/total_before*100:.2f}%)')

### Символы

import re

# Удаляем табуляции/переносы строк, цифры, знаки препинания и приводим к нижнему регистру
def clean_text(text):
    # text = re.sub(r"[A-Za-z0-9!#$%&'()*+,./:;<=>?@[\]^_`{|}~—\"\-\t\n№]+", ' ', text).lower()
    
    text = re.sub(r'[^а-яё\s]+', ' ', text.lower()) # Оставляем только кириллицу
    text = re.sub(r'\s+', ' ', text).strip()  # Убираем табуляцию/переносы и склеиваем все whitespace в один пробел
    return text

apply_and_check(df, clean_text, 'text', target_column='clean')

### Стоп-слова

from nltk.corpus import stopwords

stopwords_ru = stopwords.words("russian")

# Оставляем только "не" из всех отдельных стоп-слов для последующего соеденения после лемматизации
stopwords_ru.remove('не')

def remove_stopwords(text):
    # Если попадётся не слово или нуль
    if not isinstance(text, str) or not text.strip():
        return text
    
    words = text.split()
    
    filtered_words = [word for word in words 
                       if word and word not in stopwords_ru]
    
    return ' '.join(filtered_words)

apply_and_check(df, remove_stopwords, 'clean')

### Лишние слова по частоте

Не уверен, лучше **сначала** убрать лишние слова по частоте или проверить орфографию и сделать лемматизацию и **потом** только убрать лишние слова.  
В **первом случае** убираются все слова из списка ниже точно, но пропускаются такие же слова, если они с ошибками.  
А во **втором случае** проверка орфографии может изменить некоторые слова неправильно (что как раз и было замечено) и лемматизация может ошибочно сделать не начальную форму слов.

Также нужно убрать часто встречающиеся лишние слова по типу "счёт, банк" и подобные  
Выберем из часто встречающихся слов через value_counts(), которые могут не нести смысла для классификации

# Подсчет слов
word_counts = df['clean'].str.split().explode().value_counts()

# Вывод топ слов
word_counts.head(50)

# Выберем те, которые кажутся лишними
# Точно надо удалить их перед проверкой орфографии как можно больше, так как слишком долго обрабатывает,
# но и искать каждое слово в каждом падеже как-то неправильно. Может снова пройтись после лемматизации, так как это может убрать ещё слова
unnecessary_words = {
    'счет', 'счета', 'банк', 'банка', 'мвв', 'г', 'валюта', 'сумма',
    'сумме', 'рубль', 'курс', 'бик', 'код', 'адрес', 'россия', 'россии', 
    'филиал', 'отделение'
}

def remove_unnecessary_words(text):
    pattern = r'\b(' + '|'.join(re.escape(word) for word in unnecessary_words) + r')\b'
    return re.sub(pattern, '', text, flags=re.IGNORECASE)

apply_and_check(df, remove_unnecessary_words, 'clean')

В итоге после базовой очистки осталось ~80% символов всего датасета

### Отдельный столбец для методов выделения ключевых слов

Для этих методов нужно убрать цифры, но не пунктуацию и стоп-слова

# Отдельный класс под эти методы. Тут сложновато, так как даты, хар-ки машин, счета и другие создают множество пробелов и пустых пунктуаций
def clean_text_keyword(text, lower=True):
    # Приводим к нижнему регистру
    if lower:
        text = text.lower()

    # Убираем отдельно обозначения машин и подобные, так как они состоят не только из цифр
    text = re.sub(r'\bvin\s*[a-z0-9]{17}\b', '', text, flags=re.IGNORECASE)
    text = re.sub(r'\b(?=.*\d)(?=.*[a-z])[a-z0-9]{6,}\b', '', text, flags=re.IGNORECASE)
    
    # Убираем ВСЕ цифры
    text = re.sub(r'\d+', '', text)
    
    # Сжимаем повторяющиеся одинаковые знаки пунктуации (.,!? и т.д.)
    # "????" -> "?", ",...." -> ",."
    text = re.sub(r'([.,!?;:—\-])\1+', r'\1', text)
    
    # Убираем множественные пробелы (включая пробелы вокруг пунктуации)
    text = re.sub(r'\s+', ' ', text).strip()
    
    # Убираем пунктуацию, стоящую отдельно от слов
    text = re.sub(r'\s+([.,!?;:—\-])\s+', ' ', text)
    return text

df_keywords = df[['Id', 'text', 'answer']].copy()

apply_and_check(df_keywords, clean_text_keyword, 'text', target_column='clean_with_stopwords') 
print()

apply_and_check(df_keywords, clean_text_keyword, 'text', target_column='clean_with_stopwords_with_casing', lower=False)

# Можно через csv файл, без parquet, joblib, pickle так как датасет буквально 1000 строк
df_keywords.to_csv('data/processed/for_keywords.csv', index=False)

### Орфография

Используем Hunspell, то есть spylls, который то же самое, но новее.  
Может стоило использовать [JamSpell](https://github.com/bakwc/JamSpell), который намного быстрее и точнее, но это только на Linux, а для Windows,  версия JamSpellPro, просто не открывается сайт с инструкциями.  
Можно попробовать также C++ через обёртку Cynthon с помощью CyHunspell, но оно при установке выдаёт ошибку. Кажется оно не работает на python 3.12, только [3.10](https://pypi.org/project/cyhunspell-py310/).  
Ещё можно попробовать SymSpell, который будет поновее и быстрее во много раз со своими фичами.  
Так же попробуем PySpellChecker.

from tqdm import tqdm
import time

def print_spell_check_stats(stats, elapsed_time):
    corrected_words_percent = stats['corrected_words'] / stats['total_words'] * 100
    speed = stats['total_words'] / elapsed_time
    
    print(f'Слов проверено: {stats["total_words"]:,}')
    print(f'Исправлено слов: {stats["corrected_words"]:,} ({corrected_words_percent:.2f})%, '
          f'уникальных слов: {len(stats["unique_words"]):,}')
    print(f'Попаданий в кэш: {stats["cached_hits"]:,}')
    print(f'Скорость: {speed:.0f} слов/сек')

#### Spylls

%%script cmd /c ""
# Добавил там, где неудачные методы, чтобы пропускать ячейки, если буду нажимать "Run All Cells"

from spylls.hunspell import Dictionary

# Скачал словарик мозиллы [https://addons.mozilla.org/ru/firefox/addon/russian-spellchecking-dic-3703/]
dictionary = Dictionary.from_zip('dictionaries/russian_spellchecking_dic_3703-0.4.5.2resigned1.xpi')

word_cache = {}
stats = {
    'total_words': 0,
    'corrected_words': 0,
    'cached_hits': 0,
    'start_time': None,
}

def spell_check_spylls(text):
    # Вдруг попадётся не строка или нуль
    if not isinstance(text, str) or not text.strip():
        return text

    words = text.split()
    corrected_words = []
    
    for word in words:
        # Проверка на пустые слова или слова только из пробелов (можно убрать позже, если выявится, что слова не склеены случайно)
        if not word.strip():
            corrected_words.append(word)
            continue

        stats['total_words'] += 1
        
        # Проверяем кэш первым
        if word in word_cache:
            stats['cached_hits'] += 1
            corrected_words.append(word_cache[word])
            continue
        
        # Если нет в кэше - проверяем и кэшируем результат
        if dictionary.lookup(word):
            word_cache[word] = word
            corrected_words.append(word)
        else:
            suggestions = list(dictionary.suggest(word))
            corrected = suggestions[0] if suggestions else word

            # Если слово было исправлено
            if corrected != word:
                stats['corrected_words'] += 1
            
            word_cache[word] = corrected
            corrected_words.append(corrected)
    
    return ' '.join(corrected_words)

%%script cmd /c ""
stats['start_time'] = time.time()

tqdm.pandas(desc="Spylls")
df['corrected'] = df['clean'].progress_apply(spell_check_spylls) 
# Долго, первые 5 строк делает 6.7 секунды, без кэша ещё в 5x медленее.
# А мультипроцессорность и мультипоточность не помогает, скорее всего из-за GIL и то, что работаем в ipynb файле

elapsed_time = time.time() - stats['start_time']
print_spell_check_stats(stats, elapsed_time)

#### SymSpell

```python 
sym_spell.load_dictionary('dictionaries/ru-100k.txt', term_index=0, count_index=1)
```
Не хочет нормально открывать словарь и выдаёт слово "Рё" на каждое короткое слово, а другие не исправляет, перед этим надо исправить txt файл, так как изначально кодировка utf-8, а стандартная кодировка для моей операционной системы для python - windows-1251 и таким образом не получается прочитать файл в кодировке utf-8.  
Можно обойти через смену стандартной кодировки или пересозданием объекта словаря.

from symspellpy import SymSpell, Verbosity

entries = []
with open('dictionaries/ru-100k.txt', encoding='utf-8') as f:
    for line in f:
        parts = line.strip().split()
        entries.append((parts[0].lower(), int(parts[1])))

sym_spell = SymSpell(max_dictionary_edit_distance=2, prefix_length=7)
for word, freq in entries:
    sym_spell.create_dictionary_entry(word, freq)

word_cache = {}
stats = {
    'total_words': 0,
    'corrected_words': 0,
    'cached_hits': 0,
    'start_time': None,
    'unique_words': set() # Можно через set, чтобы не дублировать слова
}
min_len = 3 # SymSpell пытается исправить множество мелких слов, даже если это не ошибка по типу ип, пао и подобное. Ограничим

def spell_check_symspell(text):
    if not isinstance(text, str) or not text.strip():
        return text
    
    words = text.split()
    corrected_words = []

    for word in words:
        stats['total_words'] += 1

        if len(word) <= min_len:
            corrected = word
        elif word in word_cache:
            stats['cached_hits'] += 1
            corrected = word_cache[word]
        else:
            suggestions = sym_spell.lookup(word.lower(), Verbosity.CLOSEST, max_edit_distance=2, include_unknown=True)
            if suggestions:
                corrected = suggestions[0].term
            else:
                corrected = word
            word_cache[word] = corrected

        if corrected != word:
            stats['corrected_words'] += 1
            stats['unique_words'].add((word, corrected))

        corrected_words.append(corrected)
        
    return ' '.join(corrected_words)

stats['start_time'] = time.time()

tqdm.pandas(desc='SymSpell')
df['corrected'] = df['clean'].progress_apply(spell_check_symspell) 

elapsed_time = time.time() - stats['start_time']
print_spell_check_stats(stats, elapsed_time)

#### PySpellChecker

%%script cmd /c ""
from spellchecker import SpellChecker

spell_checker = SpellChecker(language='ru')

word_cache = {}
stats = {
    'total_words': 0,
    'corrected_words': 0,
    'cached_hits': 0,
    'start_time': None,
}

def spell_check_py(text):
    if not isinstance(text, str) or not text.strip():
        return text
    
    words = text.split()
    corrected_words = []
    
    for word in words:
        if not word.strip():
            corrected_words.append(word)
            continue
        
        stats['total_words'] += 1
        
        if word in word_cache:
            stats['cached_hits'] += 1
            corrected_words.append(word_cache[word])
            continue
        
        if word not in spell_checker.unknown([word]):
            word_cache[word] = word
            corrected_words.append(word)
        else:
            corrected = list(spell_checker.correction(word))[0] if spell_checker.correction(word) else word
            
            if corrected != word:
                stats['corrected_words'] += 1
            
            word_cache[word] = corrected
            corrected_words.append(corrected)
    
    return ' '.join(corrected_words)

%%script cmd /c ""
stats['start_time'] = time.time()

tqdm.pandas(desc='pyspellchecker')
df['corrected'] = df['clean'].progress_apply(spell_check_py)

elapsed_time = time.time() - stats['start_time']
print_spell_check_stats(stats, elapsed_time)

Если проверять орфографию на сыром датасете без кэша, то выйдет ~36 минут при скорости 60 слов в секунду. Примерная [таблица сравнения](https://github.com/diffitask/spell-checkers-comparison)

#### Итого исправления орфографии

Вывел сюда output, если выход ячеек сверху потеряется

```python
pyspellchecker: 100%|██████████| 1000/1000 [35:20<00:00,  2.12s/it]
Всего слов обработано: 88,529
Исправлено уникальных слов: 2,741
Попаданий в кэш: 81,487
Скорость: 42 слов/сек
```
=======================================================
```python
SymSpell: 100%|██████████| 1000/1000 [00:00<00:00, 2826.62it/s]
Слов проверено: 87,982
Исправлено слов: 8,017 (9.11)%, уникальных слов: 1,578
Попаданий в кэш: 70,319
Скорость: 245972 слов/сек
```
=======================================================
```python
Spylls: 100%|██████████| 1000/1000 [16:24<00:00,  1.02it/s]
Всего слов обработано: 88,651
Исправлено уникальных слов: 2,245
Попаданий в кэш: 81,579
Скорость: 90 слов/сек
```

Статистика по Spylls и pyspellchecker старая, так как работают 16 и 35 минут соотвественно, нету смысла повторять их из-за скорости.  
В итоге SymSpell сработал намного быстрее

### Лемматизация

# Сделаем все слова в именительном падеже с помощью лемматизации для лучшей обработки
# Также обработаем отдельно слова с "не" как "не_слово" так как много слово по типу "не удовлетворительно", так будет легче делать первичную классификацию
from pymorphy3 import MorphAnalyzer

morph = MorphAnalyzer()

def lemmatize_text(text):
    raw_tokens = text.split()

    lemmas = []
    for token in raw_tokens:
        token = token.strip()
        if token and token != 'не':
            normal_form = morph.parse(token)[0].normal_form
            lemmas.append(normal_form)
        else:
            lemmas.append(token)

    # Склеиваем "не" с последующими леммами
    result = []
    i = 0
    while i < len(lemmas):
        if lemmas[i] == 'не' and i + 1 < len(lemmas):
            # Соединяем лемму "не" со следующей леммой
            result.append('не_' + lemmas[i + 1])
            i += 2 # Пропускаем оба токена
        else:
            result.append(lemmas[i])
            i += 1
    
    return ' '.join(result)

df['lemmatized_text'] = df['corrected'].apply(lemmatize_text)

df['lemmatized_text'].iloc[5:16] # Самые короткие строки в начале датасета

# Сохраняем для других ipynb
df.to_csv('data/processed/preprocessed.csv', index=False)
